Calculate Gas transfer coefficient

Calculate Gas transfer coefficient

calc_kW(ws, wt, gas = "CO2", model = "wann14")

Arguments

ws
numeric; windspeed m/s

wt
numeric; water temperature °C

gas
character; one of 'CO2', 'CH4'

model
character; model to calculate kW. See 'Details'

Value

kW for the given parameters

Details

Currently these models are supported:

WANN14: Wanninkhof (2014)

COLE98: Cole and Caraco (1998)

References

Wanninkhof, R., 2014. Relationship between wind speed and gas exchange over the ocean revisited. Limnology and Oceanography: Methods 12, 351–362. https://doi.org/10.4319/lom.2014.12.351

Cole, J.J., Caraco, N.F., 1998. Atmospheric exchange of carbon dioxide in a low-wind oligotrophic lake measured by the addition of SF6. Limnology and Oceanography 43, 647–656. https://doi.org/10.4319/lo.1998.43.4.0647