Process GASMET data

Process GASMET data

process_gasmet(gasmet, meta, V = 0.01461, A = 0.098, pre = F, wndw = 10,
  offset = 0)

Arguments

gasmet
data frame; Recordings from GASMET.

meta
data frame; Metainformations, see 'Details'.

V
numeric; Chamber Volume. Default to 0.01461.

A
numeric; Chamber Area. Default to 0.098.

pre
if TRUE don't process the flux calculation and return the preprocessed dataframe. When FALSE (the default) data is processed with routines from 'gasfluxes' package and the fluxes are returned.

wndw
numeric; Default to 10. Number of datapoints per measurement. Can be overruled in metadata file. See 'Details'.

offset
numeric; Default to 0. Number of datapoints skipped at the beginning. Can be overruled in metadata file. See 'Details'.

Value

data frame with flux data in mmol m-2 d-1

Details

Flux calculation needs a GASMET datafile and a file containing metadata. Expected fields in the meta file are:

spot: character; ID of the measurement location.

day: character; Date of the measurement.

begin: character; Exact start of the measurement in the format HH:MM:SS. Ideally directly copied from the GASMET File.

temp: number; Temperature inside the chamber during measurement.

wndw: (optional) Number of datapoints per measurement. Function parameter is used if not provided (default: 10).

offset: (optional) Number of datapoints skipped at the beginning. Function parameter is used if not provided (default: 0).

The function splits the GASMET File into chunks containing the single chamber applications defined by begin, wndw, and offset. These chunks are passed to the gasfluxes package for the flux calculation and CO2, CH4 and N2O fluxes are returned as single data frame. gasfluxes creates a plot for every single measurement which are saved into a 'pics' subfolder within the working directory.

wndw and offset can be set as function parameters for the whole dataset or individually in the metadata file for every measurement.

Examples

# preprocessing of data (pre=TRUE) returns data frame with chunks ready to
  # pass to 'gasfluxes'.
process_gasmet(gasmet, meta_gasmet, pre=TRUE)

#>       spot        day rep       V     A        Time  CO2mmol    CH4mmol
#> 2      YT1 12.04.2017   1 0.01461 0.098 0.000000000 16.58494 0.09655897
#> 3      YT1 12.04.2017   1 0.01461 0.098 0.008333333 16.54764 0.09697339
#> 4      YT1 12.04.2017   1 0.01461 0.098 0.016666667 16.58079 0.09614456
#> 5      YT1 12.04.2017   1 0.01461 0.098 0.025000000 16.58494 0.09324364
#> 6      YT1 12.04.2017   1 0.01461 0.098 0.033333333 16.57250 0.09324364
#> 7      YT1 12.04.2017   1 0.01461 0.098 0.041666667 16.41088 0.09531572
#> 8      YT1 12.04.2017   1 0.01461 0.098 0.050000000 16.41088 0.09490131
#> 9      YT1 12.04.2017   1 0.01461 0.098 0.058333333 16.46061 0.09407247
#> 10     YT1 12.04.2017   1 0.01461 0.098 0.066666667 16.44818 0.09780222
#> 11     YT1 12.04.2017   1 0.01461 0.098 0.075000000 16.48133 0.09738780
#> 12     YT1 12.04.2017   2 0.01461 0.098 0.000000000 16.43795 0.09608477
#> 13     YT1 12.04.2017   2 0.01461 0.098 0.008055556 16.39239 0.09732725
#> 14     YT1 12.04.2017   2 0.01461 0.098 0.016388889 16.36754 0.09442814
#> 15     YT1 12.04.2017   2 0.01461 0.098 0.024722222 16.49179 0.09359982
#> 16     YT1 12.04.2017   2 0.01461 0.098 0.033055556 16.47108 0.09359982
#> 17     YT1 12.04.2017   2 0.01461 0.098 0.041388889 16.45866 0.09152903
#> 18     YT1 12.04.2017   2 0.01461 0.098 0.049722222 16.47108 0.09401398
#> 19     YT1 12.04.2017   2 0.01461 0.098 0.058055556 16.46694 0.09235734
#> 20     YT1 12.04.2017   2 0.01461 0.098 0.066388889 16.42138 0.09318566
#> 21     YT1 12.04.2017   2 0.01461 0.098 0.074722222 16.44209 0.09442814
#> 22     YT1 12.04.2017   3 0.01461 0.098 0.000000000 16.46113 0.09083140
#> 23     YT1 12.04.2017   3 0.01461 0.098 7.554444444 16.29598 0.08794131
#> 24     YT1 12.04.2017   3 0.01461 0.098 7.562777778 16.35378 0.09578584
#> 25     YT1 12.04.2017   3 0.01461 0.098 7.571111111 16.43223 0.09496010
#> 26     YT1 12.04.2017   3 0.01461 0.098 7.579444444 16.51067 0.09165714
#> 27     YT1 12.04.2017   3 0.01461 0.098 7.587777778 16.53544 0.09000566
#> 28     YT1 12.04.2017   3 0.01461 0.098 7.596111111 16.58912 0.09537297
#> 29     YT1 12.04.2017   3 0.01461 0.098 7.604444444 16.67582 0.09289575
#> 30     YT1 12.04.2017   3 0.01461 0.098 7.612777778 16.76252 0.09124427
#> 31     YT1 12.04.2017   3 0.01461 0.098 7.621111111 16.79555 0.09000566
#> 32 YTMeteo 12.04.2017   1 0.01461 0.098 0.000000000 16.32587 0.08810263
#> 33 YTMeteo 12.04.2017   1 0.01461 0.098 0.008333333 16.38378 0.09596155
#> 34 YTMeteo 12.04.2017   1 0.01461 0.098 0.016666667 16.46237 0.09513429
#> 35 YTMeteo 12.04.2017   1 0.01461 0.098 0.025000000 16.54096 0.09182527
#> 36 YTMeteo 12.04.2017   1 0.01461 0.098 0.033333333 16.56578 0.09017076
#> 37 YTMeteo 12.04.2017   1 0.01461 0.098 0.041666667 16.61955 0.09554792
#> 38 YTMeteo 12.04.2017   1 0.01461 0.098 0.050000000 16.70641 0.09306616
#> 39 YTMeteo 12.04.2017   1 0.01461 0.098 0.058333333 16.79327 0.09141165
#> 40 YTMeteo 12.04.2017   1 0.01461 0.098 0.066666667 16.82636 0.09017076
#> 41 YTMeteo 12.04.2017   1 0.01461 0.098 0.075000000 16.84291 0.09430704
#> 42 YTMeteo 12.04.2017   2 0.01461 0.098 0.000000000 16.41273 0.08810263
#> 43 YTMeteo 12.04.2017   2 0.01461 0.098 0.008055556 16.42928 0.09389341
#> 44 YTMeteo 12.04.2017   2 0.01461 0.098 0.016388889 16.50373 0.09141165
#> 45 YTMeteo 12.04.2017   2 0.01461 0.098 0.024722222 16.54923 0.09182527
#> 46 YTMeteo 12.04.2017   2 0.01461 0.098 0.033055556 16.67332 0.08975714
#> 47 YTMeteo 12.04.2017   2 0.01461 0.098 0.041388889 16.71882 0.08934351
#> 48 YTMeteo 12.04.2017   2 0.01461 0.098 0.049722222 16.75604 0.09182527
#> 49 YTMeteo 12.04.2017   2 0.01461 0.098 0.058055556 16.76432 0.08892988
#> 50 YTMeteo 12.04.2017   2 0.01461 0.098 0.066388889 16.79741 0.09017076
#> 51 YTMeteo 12.04.2017   3 0.01461 0.098 0.000000000 16.28865 0.09389341
#> 52 YTMeteo 12.04.2017   3 0.01461 0.098 0.008333333 16.40032 0.09182527
#> 53 YTMeteo 12.04.2017   3 0.01461 0.098 0.016666667 16.44996 0.09182527
#> 54 YTMeteo 12.04.2017   3 0.01461 0.098 0.025000000 16.47891 0.09099802
#> 55 YTMeteo 12.04.2017   3 0.01461 0.098 0.033333333 16.57405 0.09389341
#> 56 YTMeteo 12.04.2017   3 0.01461 0.098 0.041666667 16.65264 0.09265253
#> 57 YTMeteo 12.04.2017   3 0.01461 0.098 0.050000000 16.69400 0.09306616
#> 58 YTMeteo 12.04.2017   3 0.01461 0.098 0.058333333 16.74777 0.09430704
#> 59 YTMeteo 12.04.2017   3 0.01461 0.098 0.066666667 16.80982 0.09306616
#> 60 YTMeteo 12.04.2017   3 0.01461 0.098 0.075000000 16.85945 0.09554792
#>       N2Ommol co2 ch4 n2o
#> 2  0.01193519 co2 ch4 n2o
#> 3  0.01197663 co2 ch4 n2o
#> 4  0.01185230 co2 ch4 n2o
#> 5  0.01210095 co2 ch4 n2o
#> 6  0.01185230 co2 ch4 n2o
#> 7  0.01048473 co2 ch4 n2o
#> 8  0.01081626 co2 ch4 n2o
#> 9  0.01114780 co2 ch4 n2o
#> 10 0.01118924 co2 ch4 n2o
#> 11 0.01172798 co2 ch4 n2o
#> 12 0.01089237 co2 ch4 n2o
#> 13 0.01143077 co2 ch4 n2o
#> 14 0.01068529 co2 ch4 n2o
#> 15 0.01130653 co2 ch4 n2o
#> 16 0.01172069 co2 ch4 n2o
#> 17 0.01114086 co2 ch4 n2o
#> 18 0.01105803 co2 ch4 n2o
#> 19 0.01130653 co2 ch4 n2o
#> 20 0.01176210 co2 ch4 n2o
#> 21 0.01155502 co2 ch4 n2o
#> 22 0.01156036 co2 ch4 n2o
#> 23 0.01234481 co2 ch4 n2o
#> 24 0.01263382 co2 ch4 n2o
#> 25 0.01246867 co2 ch4 n2o
#> 26 0.01292283 co2 ch4 n2o
#> 27 0.01263382 co2 ch4 n2o
#> 28 0.01308798 co2 ch4 n2o
#> 29 0.01292283 co2 ch4 n2o
#> 30 0.01263382 co2 ch4 n2o
#> 31 0.01267511 co2 ch4 n2o
#> 32 0.01236746 co2 ch4 n2o
#> 33 0.01265700 co2 ch4 n2o
#> 34 0.01249155 co2 ch4 n2o
#> 35 0.01294654 co2 ch4 n2o
#> 36 0.01265700 co2 ch4 n2o
#> 37 0.01311199 co2 ch4 n2o
#> 38 0.01294654 co2 ch4 n2o
#> 39 0.01265700 co2 ch4 n2o
#> 40 0.01269836 co2 ch4 n2o
#> 41 0.01236746 co2 ch4 n2o
#> 42 0.01178838 co2 ch4 n2o
#> 43 0.01191247 co2 ch4 n2o
#> 44 0.01207792 co2 ch4 n2o
#> 45 0.01178838 co2 ch4 n2o
#> 46 0.01174702 co2 ch4 n2o
#> 47 0.01245018 co2 ch4 n2o
#> 48 0.01191247 co2 ch4 n2o
#> 49 0.01224337 co2 ch4 n2o
#> 50 0.01199519 co2 ch4 n2o
#> 51 0.01220201 co2 ch4 n2o
#> 52 0.01191247 co2 ch4 n2o
#> 53 0.01224337 co2 ch4 n2o
#> 54 0.01199519 co2 ch4 n2o
#> 55 0.01307062 co2 ch4 n2o
#> 56 0.01236746 co2 ch4 n2o
#> 57 0.01211928 co2 ch4 n2o
#> 58 0.01232610 co2 ch4 n2o
#> 59 0.01224337 co2 ch4 n2o
#> 60 0.01245018 co2 ch4 n2o

# prcessing provokes flux calculation for the three gas species
  # (CO2, CH4, N2O).

# NOT RUN {
process_gasmet(gasmet, meta_gasmet)
# }