Process GASMET data

Process GASMET data

process_gasmet(gasmet, meta, V = 0.01461, A = 0.098, pre = F, wndw = 10,
  offset = 0)

Arguments

gasmet
data frame; Recordings from GASMET.

meta
data frame; Metainformations, see 'Details'.

V
numeric; Chamber Volume. Default to 0.01461.

A
numeric; Chamber Area. Default to 0.098.

pre
if TRUE don't process the flux calculation and return the preprocessed dataframe. When FALSE (the default) data is processed with routines from 'gasfluxes' package and the fluxes are returned.

wndw
numeric; Default to 10. Number of datapoints per measurement. Can be overruled in metadata file. See 'Details'.

offset
numeric; Default to 0. Number of datapoints skipped at the beginning. Can be overruled in metadata file. See 'Details'.

Value

data frame with flux data in mmol m-2 d-1

Details

Flux calculation needs a GASMET datafile and a file containing metadata. Expected fields in the meta file are:

spot: character; ID of the measurement location.

day: character; Date of the measurement.

begin: character; Exact start of the measurement in the format HH:MM:SS. Ideally directly copied from the GASMET File.

temp: number; Temperature inside the chamber during measurement.

wndw: (optional) Number of datapoints per measurement. Function parameter is used if not provided (default: 10).

offset: (optional) Number of datapoints skipped at the beginning. Function parameter is used if not provided (default: 0).

The function splits the GASMET File into chunks containing the single chamber applications defined by begin, wndw, and offset. These chunks are passed to the gasfluxes package for the flux calculation and CO2, CH4 and N2O fluxes are returned as single data frame. gasfluxes creates a plot for every single measurement which are saved into a 'pics' subfolder within the working directory.

wndw and offset can be set as function parameters for the whole dataset or individually in the metadata file for every measurement.

Examples

# preprocessing of data (pre=TRUE) returns data frame with chunks ready to
  # pass to 'gasfluxes'.
process_gasmet(gasmet, meta_gasmet, pre=TRUE)

#> # A tibble: 59 x 12
#>    spot  day       rep      V     A    Time CO2mmol CH4mmol N2Ommol co2   ch4  
#>    <chr> <chr>   <dbl>  <dbl> <dbl>   <dbl>   <dbl>   <dbl>   <dbl> <chr> <chr>
#>  1 YT1   12.04.…     1 0.0146 0.098 0          16.6  0.0966  0.0119 co2   ch4  
#>  2 YT1   12.04.…     1 0.0146 0.098 0.00833    16.5  0.0970  0.0120 co2   ch4  
#>  3 YT1   12.04.…     1 0.0146 0.098 0.0167     16.6  0.0961  0.0119 co2   ch4  
#>  4 YT1   12.04.…     1 0.0146 0.098 0.025      16.6  0.0932  0.0121 co2   ch4  
#>  5 YT1   12.04.…     1 0.0146 0.098 0.0333     16.6  0.0932  0.0119 co2   ch4  
#>  6 YT1   12.04.…     1 0.0146 0.098 0.0417     16.4  0.0953  0.0105 co2   ch4  
#>  7 YT1   12.04.…     1 0.0146 0.098 0.05       16.4  0.0949  0.0108 co2   ch4  
#>  8 YT1   12.04.…     1 0.0146 0.098 0.0583     16.5  0.0941  0.0111 co2   ch4  
#>  9 YT1   12.04.…     1 0.0146 0.098 0.0667     16.4  0.0978  0.0112 co2   ch4  
#> 10 YT1   12.04.…     1 0.0146 0.098 0.075      16.5  0.0974  0.0117 co2   ch4  
#> # ... with 49 more rows, and 1 more variable: n2o <chr>

# prcessing provokes flux calculation for the three gas species
  # (CO2, CH4, N2O).

# NOT RUN {
process_gasmet(gasmet, meta_gasmet)
# }